# backend/docs/swagger.yaml
components:
  schemas:
...
    dto.CreateSubject:
      properties:
        title:
          maxLength: 255
          minLength: 1
          type: string
      required:
      - title
      type: object
    dto.CreateUser:
      properties:
        email:
          type: string
        password:
          minLength: 8
          type: string
        role:
          $ref: '#/components/schemas/repositories.UserRole'
      required:
      - email
      - password
      - role
      type: object
...
    dto.Error:
      properties:
        error:
          type: string
        message:
          type: string
      type: object
...
    dto.Subject:
      properties:
        created_at:
          type: string
        id:
          type: string
        title:
          type: string
      type: object
...
    dto.UpdateSubject:
      properties:
        title:
          maxLength: 255
          minLength: 1
          type: string
      required:
      - title
      type: object
    dto.UpdateUser:
      properties:
        email:
          type: string
        password:
          minLength: 8
          type: string
        role:
          $ref: '#/components/schemas/repositories.UserRole'
      type: object
    dto.User:
      properties:
        created_at:
          type: string
        email:
          type: string
        id:
          type: string
        role:
          $ref: '#/components/schemas/repositories.UserRole'
        updated_at:
          type: string
      type: object
...
    repositories.UserRole:
      enum:
      - administrator
      - teacher
      - student
      - administrator
      - teacher
      - student
      - administrator
      - teacher
      - student
      type: string
      x-enum-varnames:
      - UserRoleAdministrator
      - UserRoleTeacher
      - UserRoleStudent
  securitySchemes:
    BearerAuth:
      in: header
      name: Authorization
      type: apiKey
...
info:
  description: REST API образовательной системы Edutrack.
  title: Edutrack API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /subjects:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/dto.Subject'
                type: array
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Получить список предметов
      tags:
      - subjects
    post:
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/dto.CreateSubject'
                description: Данные предмета
                summary: request
        description: Данные предмета
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Subject'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Создать предмет
      tags:
      - subjects
  /subjects/{id}:
    delete:
      parameters:
      - description: Идентификатор предмета
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Удалить предмет
      tags:
      - subjects
    get:
      parameters:
      - description: Идентификатор предмета
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Subject'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Получить предмет по идентификатору
      tags:
      - subjects
    patch:
      parameters:
      - description: Идентификатор предмета
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/dto.UpdateSubject'
                description: Данные предмета
                summary: request
        description: Данные предмета
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Subject'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Обновить предмет
      tags:
      - subjects
  /users:
    post:
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/dto.CreateUser'
                description: Данные пользователя
                summary: request
        description: Данные пользователя
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.User'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Forbidden
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Создать пользователя
      tags:
      - users
  /users/{id}:
    delete:
      parameters:
      - description: Идентификатор пользователя
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Удалить пользователя
      tags:
      - users
    get:
      parameters:
      - description: Идентификатор пользователя
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.User'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Получить пользователя по идентификатору
      tags:
      - users
    patch:
      parameters:
      - description: Идентификатор пользователя
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/dto.UpdateUser'
                description: Данные пользователя
                summary: request
        description: Данные пользователя
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.User'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Частично обновить пользователя
      tags:
      - users
servers:
- url: /api